requirements.txt
streamlit>=1.35.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
cantools>=39.0.0
pypdf>=4.0.0
pandas>=2.2.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
sqlalchemy>=2.0.0
